Functionality : Initiated frontend modules and connected Chakra UI , minor port changes

+++ mern_crash_course/backend/controllers/product.controller.js
@@ -53,12 +53,17 @@ export const updatedProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   //deletion api end point
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid product Id" });
+  }
   try {
     await Product.findByIdAndDelete(id);
     res
       .status(200)
       .json({ success: true, message: "Product deleted successfully" });
   } catch (error) {
-    res.status(404).json({ success: false, message: "Product not found" });
+    res.status(500).json({ success: false, message: "Sever Error" });
   }
 };
